# src/swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/user": {
      "get": {
        "operationId": "UserController_paginateAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGeneralReponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCreateDUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/address/find": {
      "get": {
        "operationId": "UserController_getAddress",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/address": {
      "post": {
        "operationId": "UserController_createAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
    "/category": {
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "category"
        ]
      },
      "get": {
        "operationId": "CategoryController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGeneralReponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "category"
        ]
      }
    },
...
    "/ticket": {
      "post": {
        "operationId": "TicketController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ticket"
        ]
      }
    },
    "/ticket/history": {
      "get": {
        "operationId": "TicketController_getUserTickets",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGeneralReponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ticket"
        ]
      }
    },
    "/ticket/admin": {
      "get": {
        "operationId": "TicketController_getAllTickets",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ticket"
        ]
      }
    },
    "/ticket/{id}": {
      "get": {
        "operationId": "TicketController_findOneById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ticket"
        ]
      }
    },
...
  },
  "info": {
    "title": "klickum-api-doc",
    "description": "klickum-api ecommerce API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "MetaDataDto": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "number"
          },
          "itemCount": {
            "type": "number"
          },
          "itemsPerPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          }
        },
        "required": [
          "totalItems",
          "itemCount",
          "itemsPerPage",
          "totalPages",
          "currentPage"
        ]
      },
      "PaginatedGeneralReponseDto": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/MetaDataDto"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "meta",
          "items"
        ]
      },
      "UpdateCreateDUserDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "balance": {
            "type": "number"
          }
        },
        "required": [
          "isActive",
          "balance"
        ]
      },
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "zipcode": {
            "type": "string"
          },
          "streetNumber": {
            "type": "string"
          },
          "streetName": {
            "type": "string"
          },
          "countryId": {
            "type": "number"
          },
          "cityName": {
            "type": "string"
          }
        },
        "required": [
          "zipcode",
          "streetNumber",
          "streetName",
          "countryId",
          "cityName"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "raffleId": {
            "type": "string"
          },
          "code": {
            "type": "number"
          }
        },
        "required": [
          "raffleId",
          "code"
        ]
      },
...
    }
  }
}